+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/roomlist/header/AccountData.kt
@@ -38,7 +38,9 @@ class AccountDataViewImpl : AccountDataView {
                     verticalAlignment = Alignment.CenterVertically
                 ) {
                     AccountAvatar(accountViewModel)
-                    CloseProfile(roomListViewModel)
+                    if(roomListViewModel.closeProfileNeeded){
+                        CloseProfile({roomListViewModel.closeProfile()})
+                    }
                     ShowSearch(roomListViewModel)
                     AccountOptions(accountViewModel, roomListViewModel)
                 }