+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/roomlist/header/CloseProfile.kt
@@ -13,32 +13,29 @@ import de.connect2x.trixnity.messenger.compose.view.get
 import de.connect2x.trixnity.messenger.compose.view.i18n.I18nView
 import de.connect2x.trixnity.messenger.compose.view.theme.components
 import de.connect2x.trixnity.messenger.compose.view.theme.components.ThemedIconButton
-import de.connect2x.trixnity.messenger.viewmodel.roomlist.RoomListViewModel
 
 interface CloseProfileView {
     @Composable
-    fun create(roomListViewModel: RoomListViewModel)
+    fun create(closeProfile: () -> Unit)
 }
 
 @Composable
-fun CloseProfile(roomListViewModel: RoomListViewModel) {
-    DI.get<CloseProfileView>().create(roomListViewModel)
+fun CloseProfile(closeProfile: () -> Unit) {
+    DI.get<CloseProfileView>().create(closeProfile)
 }
 
 class CloseProfileViewImpl : CloseProfileView {
     @Composable
-    override fun create(roomListViewModel: RoomListViewModel) {
+    override fun create(closeProfile: () -> Unit) {
         val i18n = DI.get<I18nView>()
 
-        if (roomListViewModel.closeProfileNeeded) {
-            Box {
-                Tooltip({ Text(i18n.accountCloseProfile()) }) {
-                    ThemedIconButton(
-                        style = MaterialTheme.components.destructiveIconButton,
-                        onClick = { roomListViewModel.closeProfile() },
-                    ) {
-                        Icon(Icons.AutoMirrored.Filled.Logout, i18n.accountCloseProfile())
-                    }
+        Box {
+            Tooltip({ Text(i18n.accountCloseProfile()) }) {
+                ThemedIconButton(
+                    style = MaterialTheme.components.destructiveIconButton,
+                    onClick = { closeProfile() },
+                ) {
+                    Icon(Icons.AutoMirrored.Filled.Logout, i18n.accountCloseProfile())
                 }
             }
         }