+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/settings/ProfilesSettings.kt
@@ -7,6 +7,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
@@ -15,11 +16,13 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import de.connect2x.trixnity.messenger.compose.view.DI
 import de.connect2x.trixnity.messenger.compose.view.VerticalScrollbar
-import de.connect2x.trixnity.messenger.compose.view.common.ApprovableTextField
 import de.connect2x.trixnity.messenger.compose.view.common.Header
 import de.connect2x.trixnity.messenger.compose.view.common.RunningText
-import de.connect2x.trixnity.messenger.compose.view.i18n.I18nView
+import de.connect2x.trixnity.messenger.compose.view.common.SmallSpacer
 import de.connect2x.trixnity.messenger.compose.view.get
+import de.connect2x.trixnity.messenger.compose.view.i18n.I18nView
+import de.connect2x.trixnity.messenger.compose.view.theme.components
+import de.connect2x.trixnity.messenger.compose.view.theme.components.ThemedButton
 import de.connect2x.trixnity.messenger.compose.view.theme.components.ThemedListItemSwitch
 import de.connect2x.trixnity.messenger.viewmodel.settings.ProfilesSettingsViewModel
 
@@ -38,39 +41,66 @@ class ProfilesSettingsViewImpl : ProfilesSettingsView {
     override fun create(profilesSettingsViewModel: ProfilesSettingsViewModel) {
         val i18n = DI.get<I18nView>()
         val scroll = rememberScrollState()
-        val multiProfileEnabled = profilesSettingsViewModel.isMultiProfile.collectAsState().value
-        val canChangeMultiProfileMode = profilesSettingsViewModel.canChangeMultiProfileMode.collectAsState().value
+        val profilesDialogueController = ProfilesDialogueController()
 
         Box(Modifier.fillMaxSize()) {
             Column {
                 Header(profilesSettingsViewModel::close, i18n.profilesSettings())
-
                 Box {
                     Column(Modifier.padding(10.dp).verticalScroll(scroll)) {
-                        SettingsCard(i18n.multiProfilesModeSettings()) {
-                            RunningText(i18n.multiProfileModeDescription())
-                                ThemedListItemSwitch(
-                                    headlineContent = { Text(i18n.profileSelectionMultipleAccountSwitch()) },
-                                    enabled = canChangeMultiProfileMode,
-                                    selected = multiProfileEnabled,
-                                    supportingContent = {if(!canChangeMultiProfileMode){Text(i18n.cannotDisableMultiProfileMode())} },
-                                    onChange = {profilesSettingsViewModel.setMultiProfileEnabled(it)}
-                                )
-                        }
-                        SettingsCard(i18n.profileNameSettings()) {
-                            ApprovableTextField(
-                                profilesSettingsViewModel.profileNameTextFieldViewModel,
-                                true,
-                                Modifier,
-                                i18n.profileNameTextfield(),
-                                i18n.profileNamePlaceholder(),
-                            )
-                        }
+                        MultiProfileSettingsCard(profilesSettingsViewModel)
+                        ProfileListSettingsCard(profilesSettingsViewModel, profilesDialogueController)
                     }
                     VerticalScrollbar(
                         Modifier.align(Alignment.CenterEnd).fillMaxHeight(),
                         scroll,
                     )
+                }
+            }
+        }
+    }
+
+    @Composable
+    private fun MultiProfileSettingsCard(profilesSettingsViewModel: ProfilesSettingsViewModel) {
+        val i18n = DI.get<I18nView>()
+        val multiProfileEnabled = profilesSettingsViewModel.isMultiProfile.collectAsState().value
+        val canChangeMultiProfileMode = profilesSettingsViewModel.canChangeMultiProfileMode.collectAsState().value
+
+        SettingsCard(i18n.multiProfilesModeSettings()) {
+            RunningText(i18n.multiProfileModeDescription())
+            ThemedListItemSwitch(
+                headlineContent = { Text(i18n.profileSelectionMultipleAccountSwitch()) },
+                enabled = canChangeMultiProfileMode,
+                selected = multiProfileEnabled,
+                supportingContent = {
+                    if (!canChangeMultiProfileMode) {
+                        Text(i18n.cannotDisableMultiProfileMode())
+                    }
+                },
+                onChange = { profilesSettingsViewModel.setMultiProfileEnabled(it) }
+            )
+        }
+    }
+
+    @Composable
+    private fun ProfileListSettingsCard(
+        profilesSettingsViewModel: ProfilesSettingsViewModel,
+        profilesDialogueController: ProfilesDialogueController
+    ) {
+        val i18n = DI.get<I18nView>()
+        val multiProfileEnabled = profilesSettingsViewModel.isMultiProfile.collectAsState().value
+
+        SettingsCard(i18n.profilesSettingsList()) {
+            ProfileList(profilesSettingsViewModel, profilesDialogueController)
+            SmallSpacer()
+            if (multiProfileEnabled) {
+                ThemedButton(
+                    style = MaterialTheme.components.primaryButton,
+                    onClick = {
+                        profilesDialogueController.openCreateDialogue(profilesSettingsViewModel.activeProfile.value)
+                    },
+                ) {
+                    Text(i18n.selectProfileCreateInstead())
                 }
             }
         }